feat(games): verify independent Falcon scheduling through SQL publication stalls
Drive the existing fixed-step runtime from a Rust worker clock while Godot pauses consumption across rollback. Exercise three pinned SQL slots, skipped publication, corrected-window recovery, full-state equality, and exact mesh readback. Include the reproducible MP4, audit traces, and scope documentation.

Directed by: Chris Hafley approved committing the incremental core and proving independent schedules with renderer stalls.
Implemented by: Codex implemented the scheduling fixture, SQL checks, Godot capture, and recorded evidence.

+++ games/blender-godot-sqlite-proof/simulation-core/1_sandbag.rs
@@ -98,9 +98,9 @@ impl Sandbag {
         self.knockback = calc::knockback(
             hit.damage,
             hit.damage,
-            hit.kbg as u32,
-            hit.bkb as u32,
-            hit.wdsk as u32,
+            hit.kbg,
+            hit.bkb,
+            hit.wdsk,
             false,
             &target,
             percent_before,
@@ -111,11 +111,7 @@ impl Sandbag {
             false,
             false,
         );
-        let angle = calc::resolve_sakurai_angle(
-            (hit.trajectory as f32).to_radians(),
-            self.knockback,
-            false,
-        );
+        let angle = calc::resolve_sakurai_angle(hit.trajectory.to_radians(), self.knockback, false);
         self.velocity = [
             0.0,
             calc::initial_y_velocity(self.knockback, angle, false),
